+++ aggregation_mode/proof_aggregator/src/bin/proof_aggregator_cpu.rs
@@ -1,0 +1,4 @@
+#[tokio::main]
+async fn main() {
+    proof_aggregator::run().await;
+}
